327107 - remove orphan filter sliders from the UI
when upgrading to bs5 some filters were duplicated and left as orphans in the table body.

The options fillContainer and scrollX seems to be the culprits. Although I could not find specific issues on this there seems to be many threads on the interactions of these two options with other datatable features.

The included solution is to remove these elements by hand on each draw.